# Optimize labels.String()