--- a/pkg/config/labels/instance.go
+++ b/pkg/config/labels/instance.go
@@ -15,14 +15,14 @@
 package labels
 
 import (
-	"bytes"
 	"fmt"
 	"regexp"
-	"sort"
+	"strings"
 
 	"github.com/hashicorp/go-multierror"
 
 	"istio.io/istio/pkg/maps"
+	"istio.io/istio/pkg/slices"
 )
 
 const (
@@ -140,25 +140,25 @@ func validateTagKey(k string) error {
 }
 
 func (i Instance) String() string {
-	labels := make([]string, 0, len(i))
-	for k, v := range i {
-		if len(v) > 0 {
-			labels = append(labels, fmt.Sprintf("%s=%s", k, v))
-		} else {
-			labels = append(labels, k)
-		}
-	}
-	sort.Strings(labels)
+	// Ensure stable ordering
+	keys := slices.Sort(maps.Keys(i))
 
-	var buffer bytes.Buffer
+	var buffer strings.Builder
+	// Assume each kv pair is roughly 10 characters. We could be under or over, this is just a guess to optimize
+	buffer.Grow(len(keys) * 10)
 	first := true
-	for _, label := range labels {
+	for _, k := range keys {
+		v := i[k]
 		if !first {
 			buffer.WriteString(",")
 		} else {
 			first = false
 		}
-		buffer.WriteString(label)
+		if len(v) > 0 {
+			buffer.WriteString(k + "=" + v)
+		} else {
+			buffer.WriteString(k)
+		}
 	}
 	return buffer.String()
 }