--- a/pkg/config/labels/instance_test.go
+++ b/pkg/config/labels/instance_test.go
@@ -15,7 +15,6 @@
 package labels_test
 
 import (
-	"fmt"
 	"testing"
 
 	"istio.io/istio/pkg/config/labels"
@@ -203,7 +202,7 @@ func TestInstanceValidate(t *testing.T) {
 func BenchmarkLabelString(b *testing.B) {
 	big := labels.Instance{}
 	for i := 0; i < 50; i++ {
-		big[fmt.Sprint(i)] = "some value"
+		big["topology.kubernetes.io/region"] = "some value"
 	}
 	small := labels.Instance{
 		"app": "foo",