```
name                 old time/op    new time/op    delta
LabelString/small-8     481ns ± 4%     169ns ± 1%  -64.96%  (p=0.008 n=5+5)
LabelString/big-8      12.2µs ± 3%     4.9µs ± 3%  -59.29%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
LabelString/small-8      216B ± 0%       56B ± 0%  -74.07%  (p=0.008 n=5+5)
LabelString/big-8      6.01kB ± 0%    2.30kB ± 0%  -61.67%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
LabelString/small-8      10.0 ± 0%       2.0 ± 0%  -80.00%  (p=0.008 n=5+5)
LabelString/big-8         158 ± 0%         3 ± 0%  -98.10%  (p=0.008 n=5+5)
```

This is shown to be a hot path in some scenarios (https://github.com/istio/istio/issues/44985)