@howardjohn 2023-05-31
This is less important since https://github.com/istio/istio/pull/45230 removes it from the hotpath. But might as well have it more efficient